# src/main/java/com/example/testjob/api/api.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
paths:
  /departure-mailing:
    post:
      tags:
        - mailing-controller
      summary: departureMailing
      operationId: departureMailingUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: index
          in: query
          description: index - индекс почтового отделения из которого отправляется посылка
          required: true
          type: integer
          format: int32
        - name: trackingNumber
          in: query
          description: trackingNumber - трек код почтового отправления
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: BAD REQUEST - если указаны не все параметры, то приходит сообщение с указанием какого параметра не хватает
          schema:
            $ref: '#/definitions/Message'
        '500':
          description: INTERNAL SERVER ERROR
          schema:
            $ref: '#/definitions/Message'
  /end-mailing:
    patch:
      tags:
        - mailing-controller
      summary: endMailing
      operationId: endMailingUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: index
          in: query
          description: index - индекс почтового отделения из которого отправляется посылка
          required: true
          type: integer
          format: int32
        - name: trackingNumber
          in: query
          description: trackingNumber - трек код почтового отправления
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: BAD REQUEST - если указаны не все параметры, то приходит сообщение с указанием какого параметра не хватает
          schema:
            $ref: '#/definitions/Message'
        '500':
          description: INTERNAL SERVER ERROR
          schema:
            $ref: '#/definitions/Message'

  /mailing:
    get:
      tags:
        - mailing-controller
      summary: getMailing
      operationId: getMailingUsingGET
      produces:
        - '*/*'
      parameters:
        - name: trackingNumber
          in: query
          description: trackingNumber - трек код почтового отправления
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: BAD REQUEST - если указаны не все параметры, то приходит сообщение с указанием какого параметра не хватает
          schema:
            $ref: '#/definitions/Message'
        '500':
          description: INTERNAL SERVER ERROR
          schema:
            $ref: '#/definitions/Message'
    post:
      tags:
        - mailing-controller
      summary: createMailing
      operationId: createMailingUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: index
          in: query
          description: Индекс почтового отделения, где создается посылка
          required: true
          type: integer
          format: int32
        - in: body
          name: mailing
          description: Почтовое отправление, обязательные поля-type,index,endAddress,recipientName. Поле trackingNumber генерируется с помощь хеш кода на основание вышеперечисленных полей. Поле indexPostOffice передает индекс почтвового отделения где регистрируют посылку
          required: true
          schema:
            $ref: '#/definitions/CreateMailing'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: BAD REQUEST - если указаны не все параметры, то приходит сообщение с указанием какого параметра не хватает
          schema:
            $ref: '#/definitions/Message'
        '500':
          description: INTERNAL SERVER ERROR
          schema:
            $ref: '#/definitions/Message'

  /post-office:
    post:
      tags:
        - post-office-controller
      summary: createPostOffice
      operationId: createPostOfficeUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: postOffice
          description: Почтовое отправление, обязательные поля - index,namePostOffice,address
          required: true
          schema:
            $ref: '#/definitions/PostOffice'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: BAD REQUEST - если указаны не все параметры, то приходит сообщение с указанием какого параметра не хватает
          schema:
            $ref: '#/definitions/Message'
        '500':
          description: INTERNAL SERVER ERROR
          schema:
            $ref: '#/definitions/Message'
  /receiving-mailing:
    post:
      tags:
        - mailing-controller
      summary: receivingMailing
      operationId: receivingMailingUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: index
          in: query
          description: index
          required: true
          type: integer
          format: int32
        - name: trackingNumber
          in: query
          description: trackingNumber - трек код почтового отправления
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: BAD REQUEST - если указаны не все параметры, то приходит сообщение с указанием какого параметра не хватает
          schema:
            $ref: '#/definitions/Message'
        '500':
          description: INTERNAL SERVER ERROR
          schema:
            $ref: '#/definitions/Message'
definitions:
  Message:
    type: object
    properties:
      message:
        type: string
  IntermediatePoint:
    type: object
    properties:
      dateOfArrival:
        type: string
      dateOfDeparture:
        type: string
      id:
        type: integer
        format: int64
      indexPostOffice:
        type: integer
        format: int32
    title: IntermediatePoint
  Mailing:
    type: object
    properties:
      endAddress:
        type: string
      id:
        type: integer
        format: int64
      index:
        type: integer
        format: int32
      listPostOffices:
        type: array
        items:
          $ref: '#/definitions/IntermediatePoint'
      recipientName:
        type: string
      status:
        type: boolean
      trackingNumber:
        type: integer
        format: int32
      type:
        type: integer
        format: int32
    title: Mailing

  CreateMailing:
    type: object
    properties:
      type:
        type: integer
        format: int32
      index:
        type: integer
        format: int32
      endAddress:
        type: string
      recipientName:
        type: string
      indexPostOffice:
        type: integer
        format: int32

  PostOffice:
    type: object
    properties:
      address:
        type: string
      index:
        type: integer
        format: int32
      namePostOffice:
        type: string
    title: PostOffice